Many miscellaneous changes
Updated the readme to fix a few typos and make it more detailed.

Finished fixing the Macaw Masked skins to have the correct rimlight for each weapon and to not have a very hard to notice and useless reflection. cubemap.

Did a few touch-ups of lightwarps for the wearable weapons.

Gave a $detail for fire for every wearable cosmetic that missed it.

Fixed the Detonator and the Sharp Dresser missing rimlight.