Fixed a few cosmetics
Fixed self-illumination for the Crone's Dome not working.
Fixed Something Special for Someone Special to both prevent console spam and work properly in DX80.
Added missing rimlight to the Treasure and Bounty hats, they've also been fixed to not be completely reflective on DX80.
Fixed the Spirit of Giving not cloaking properly.