Many changes and fixes
Fixed the Sentry Buster to always have BLU legs in some custom MvM missions.
Fixed the Concheror's backpack to have the fire effect when the Soldier is burning.
Fixed the Climbing Commander to not be fully reflective on DX80.
Fixed the BLU variant of Mr. Juice missing the normal map.
Updated the Saxxy with fixes from Rabscootle (which are already in vanilla)
Fixed the BLU variant of the Private Eye to not have cubemap reflectivity.
Restored Doe-Boy's vanilla weapon lightwarp.
Fixed Jarate giving neon colors to the Argyle Ace, gave it a consistent lightwarp.
Engineer Bots no longer have their whole head glowing on DX80.